25.09.29
스킬 드래그시 스킬 설명 출력
스킬 조합시 데미지 미적용 버그 수정

+++ Assets/Scripts/BuildSetting/CharacterSaveData.cs
@@ -9,7 +9,7 @@ public class CharacterSaveData
     public int Player_level;
     public int Player_currentEXP;
     public int Player_HP;
-    public int Player_mana;
+    public float Player_mana;
     public int Player_gold;
     public int Player_stat;
     public int Player_int;
